refactor: move sub-components to components.rs (#165)
